tests/test_query_manager.py
import pytest
from data_quality_framework.query_manager import QueryManager

# You might also need to import any other necessary classes or methods.


class TestDDL:
    """
    Test cases related to DDL operations.
    """

    @pytest.fixture
    def setup_query_manager(self):
        """
        Setup fixture for QueryManager instance.
        """
        qm = QueryManager(db_type="snowflake")  # Example initialization
        return qm

    def test_compare_ddl_snowflake_mysql(self, setup_query_manager):
        """
        Test if DDL comparison between Snowflake and MySQL works as expected.
        """
        qm = setup_query_manager
        result = qm.compare_ddl(table_name="sample_table")
        # Assuming your method returns True for successful comparison, otherwise False
        assert result is True, "Comparison of DDL between Snowflake and MySQL failed."

    def test_fetch_ddl_snowflake(self, setup_query_manager):
        """
        Test fetching DDL from Snowflake.
        """
        qm = setup_query_manager
        ddl = qm.fetch_ddl(db_type="snowflake", table_name="sample_table")
        assert ddl is not None, "Failed to fetch DDL from Snowflake."

    def test_fetch_ddl_mysql(self, setup_query_manager):
        """
        Test fetching DDL from MySQL.
        """
        qm = setup_query_manager
        ddl = qm.fetch_ddl(db_type="mysql", table_name="sample_table")
        assert ddl is not None, "Failed to fetch DDL from MySQL."


class TestDML:
    """
    Test cases related to DML operations.
    """

    def test_select_snowflake(self, setup_query_manager):
        """
        Test SELECT query execution on Snowflake.
        """
        qm = setup_query_manager
        result = qm.execute_select(db_type="snowflake", query="SELECT * FROM sample_table LIMIT 10")
        assert result is not None and len(result) == 10, "Failed to execute SELECT on Snowflake."

    def test_select_mysql(self, setup_query_manager):
        """
        Test SELECT query execution on MySQL.
        """
        qm = setup_query_manager
        result = qm.execute_select(db_type="mysql", query="SELECT * FROM sample_table LIMIT 10")
        assert result is not None and len(result) == 10, "Failed to execute SELECT on MySQL."

    def test_compare_dml_snowflake_mysql(self, setup_query_manager):
        """
        Test data comparison between Snowflake and MySQL after executing SELECT queries.
        """
        qm = setup_query_manager
        result = qm.compare_dml(query="SELECT * FROM sample_table LIMIT 10")
        # Assuming your method returns True for successful comparison, otherwise False
        assert result is True, "Comparison of data between Snowflake and MySQL after SELECT query failed."

